+++ include/nfx/Gl.h
@@ -8,5 +8,6 @@
  */
 
 #include "nfx/graphics/gl/core/buffers/Buffer.h"
+#include "nfx/graphics/gl/core/textures/Texture2D.h"
 #include "nfx/graphics/gl/core/Context.h"
 #include "nfx/graphics/gl/core/GlDefinitions.h"